requirement.txt
Django==6.0
djangorestframework
psycopg2-binary
drf_spectacular